[trivial] Adjusted CPU clock speed to match cycles per frame.

+++ SIDFactoryII/source/runtime/execution/executionhandler.cpp
@@ -46,7 +46,7 @@ namespace Emulation
 		, m_CPUFrameCounter(0)
 		, m_UpdateEnabled(false)
 	{
-		m_CyclesPerFrame = static_cast<unsigned int>(EMULATION_CYCLES_PER_FRAME_PAL);
+		m_CyclesPerFrame = EMULATION_CYCLES_PER_FRAME_PAL;
 
 		// Create a sample buffer. The sample frequency is used for determining the size, which is probably 50 times the size required.
 		m_SampleBufferSize = (static_cast<unsigned int>(pSIDProxy->GetSampleFrequency()) << 8);
@@ -201,7 +201,7 @@ namespace Emulation
 
 	void ExecutionHandler::SetPAL(const bool inPALMode)
 	{
-		m_CyclesPerFrame = static_cast<unsigned int>(inPALMode ? EMULATION_CYCLES_PER_FRAME_PAL : EMULATION_CYCLES_PER_FRAME_NTSC);
+		m_CyclesPerFrame = inPALMode ? EMULATION_CYCLES_PER_FRAME_PAL : EMULATION_CYCLES_PER_FRAME_NTSC;
 	}
